7-Zip Version 21.03
What's new after 7-Zip 21.02 alpha:
- The maximum dictionary size for LZMA/LZMA2 compressing was increased to 4 GB (3840 MiB).
- Minor speed optimizations in LZMA/LZMA2 compressing.
- see: https://sourceforge.net/p/sevenzip/discussion/45797/thread/9f5b067368/

+++ CPP/7zip/Archive/7z/7zHandlerOut.cpp
@@ -133,7 +133,7 @@ HRESULT CHandler::SetMainMethod(
     if (_numSolidBytesDefined)
       continue;
 
-    UInt32 dicSize;
+    UInt64 dicSize;
     switch (methodFull.Id)
     {
       case k_LZMA:
